requirements.txt
pycountry~=22.3.5
setuptools~=60.2.0
Pillow~=9.4.0
fuzzywuzzy~=0.18.0
pixelmatch~=0.3.0
python-Levenshtein